Remove old method of assuming role
